cancel_timer_2 in same thread returns ok
