+++ lumen_runtime/src/otp/erlang/cancel_timer_2/test/with_reference_timer_reference/with_list_options/with_async_false/with_info_false/with_local_reference/with_timer/in_same_thread.rs
@@ -1,58 +1,16 @@
 use super::*;
 
-use crate::test::{timeout_after_half, with_timer_in_same_thread};
-
 #[test]
 #[ignore]
 fn without_timeout_returns_ok_and_does_not_send_timeout_message() {
-    with_timer_in_same_thread(|milliseconds, message, timer_reference, process| {
-        timeout_after_half(milliseconds);
-
-        let timeout_message = timeout_message(timer_reference, message, process);
-
-        assert!(!has_message(process, timeout_message));
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-
-        // again before timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-
-        timeout_after_half(milliseconds);
-
-        assert!(!has_message(process, timeout_message));
-
-        // again after timeout
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-    })
+    super::with_timer_in_same_thread_without_timeout_returns_ok_and_does_not_send_timeout_message(
+        options,
+    );
 }
 
 #[test]
 fn with_timeout_returns_ok_after_timeout_message_was_sent() {
-    with_timer_in_same_thread(|milliseconds, message, timer_reference, process| {
-        timeout_after(milliseconds);
-
-        let timeout_message = timeout_message(timer_reference, message, process);
-
-        assert_has_message!(process, timeout_message);
-
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-
-        // again
-        assert_eq!(
-            native(process, timer_reference, options(process)),
-            Ok(Atom::str_to_term("ok"))
-        );
-    })
+    super::with_timer_in_same_thread_with_timeout_returns_ok_after_timeout_message_was_sent(
+        options,
+    );
 }
